+++ src/sxhkd.h
@@ -60,7 +60,7 @@ extern int redir_fd;
 extern FILE *status_fifo;
 extern char progress[3 * MAXLEN];
 extern int mapping_count;
-extern int timeout;
+extern float timeout;
 
 extern hotkey_t *hotkeys_head, *hotkeys_tail;
 extern bool running, grabbed, toggle_grab, reload, bell, chained, locked;
